+++ tests/Unit/PdfTest.php
@@ -167,6 +167,10 @@ class PdfTest extends TestCase
                 ['no-http-redirects' => true, 'fail-on-http-errors' => true, 'verbose' => true, 'debug' => true, 'info' => true, 'version' => true],
                 '/' . $q . 'emptyBinary' . $q . ' --timeout \d* --info --verbose --debug --version --no-http-redirects --fail-on-http-errors ' . $q . '.*\.html' . $q . ' ' . $q . '.*\.pdf' . $q . '/',
             ],
+            '10 - output intent option' => [
+                ['output-intent' => 'device-cmyk'],
+                '/' . $q . 'emptyBinary' . $q . ' --timeout \d* --output-intent ' . $q . 'device-cmyk' . $q . ' ' . $q . '.*\.html' . $q . ' ' . $q . '.*\.pdf' . $q . '/',
+            ],
         ];
     }